Create 0_Perceptron_Logic_Gates.py
In this project, we will use perceptrons to model the fundamental building blocks of computers — logic gates.
Create dataset of four points and labels of AND gate.
Let’s plot these four points on a graph.
Now that we have our data, let’s build a perceptron to learn AND. Create a Perceptron object named classifier. For now, set the parameter max_iter to 40.
We’ll now train the model.
Let’s see if the algorithm learned AND. Note that it is pretty unusual to train and test on the same dataset. In this case, since there are only four possible inputs to AND, we’re stuck training on every possible input and testing on those same points.
Your perceptron should have 100% accuracy! You just taught it an AND gate!
Change labels so your data represents XOR, OR gate.
We know the perceptron has been trained correctly, but let’s try to visualize what decision boundary it is making.
Try calling classifier‘s .decision_function() method using [[0, 0], [1, 1], [0.5, 0.5]] as a parameter. Print the results.
If we do this for a grid of points, we can make a heat map that reveals the decision boundary.