Update 2_Life_Expectancy.py
Implementing Neural Networks
The World Health Organization (WHO)’s Global Health Observatory (GHO) data repository tracks life expectancy for countries worldwide by following health status and many other related factors.

Although there have been a lot of studies undertaken in the past on factors affecting life expectancy considering demographic variables, income composition, and mortality rates, it was found that the effects of immunization and human development index were not taken into account.

This dataset covers a variety of indicators for all countries from 2000 to 2015 including:

immunization factors
mortality factors
economic factors
social factors
other health-related factors
Ideally, this data will eventually inform countries concerning which factors to change in order to improve the life expectancy of their populations. If we can predict life expectancy well given all the factors, this is a good sign that there are some important patterns in the data. Life expectancy is expressed in years, and hence it is a number. This means that in order to build a predictive model one needs to use regression.

In this project, you will design, train, and evaluate a neural network model performing the task of regression to predict the life expectancy of countries using this dataset. Excited? Let’s go!

Load the life_expectancy.csv dataset into a pandas DataFrame.
Observe the data by printing the first five entries in the DataFrame dataset.
Drop the Country column from the DataFrame using the DataFrame drop method. Why? To create a predictive model, knowing from which country data comes can be confusing and it is not a column we can generalize over. We want to learn a general pattern for all the countries, and not only those dependent on specific countries.
In the next two steps, you will split the data into labels and features.
Features span from the first column to the last column (not including it, since it’s a label column in our dataset). Create a new variable called features.
When you observed your dataset you probably noticed that some columns are categorical. We learned in this lesson that categorical columns need to be converted into numerical columns.
Split your data into training set and test sets.
The next step is to standardize/normalize your numerical features.
Fit your instance ct of ColumnTransformer to the training data and at the same time transform it.
Transform your test data instance features_test.
Create an instance of the tensorflow.keras.models.Sequential model called my_model.
Create and add the input layer to the network model.
Add one hidden layer with any number of hidden units you wish.
Add a output layer with one neuron since you need a single output for a regression prediction.
Print the summary of the model.
Create an instance of the Adam optimizer with the learning rate equal to 0.01.
Once your optimizer is initialized, compile the model. 
Assign the following parameters:
For loss use the Mean Squared Error (mse)
For metrics use the Mean Absolute Error (mae)
For opt (the optimizer parameters) use the instance of the optimizer you created in the previous step.

Train your model with the fit() method by passing it the following parameters:
your preprocessed training data
training labels
number of epochs equal to 40
batch size equal to 1
verbose set to 1

Evaluate your trained model.
Print your final loss (RMSE) and final metric (MAE) to check the predictive performance on the test set.